+++ client/src/config/placeSearchTags.ts
@@ -93,3 +93,10 @@ export const relatedAccordions: IHalLinks = {
   ...relatedTypes,
   locations,
 }
+
+// Used for rendering hierarchy children
+export const hierarchyChildren: IHalLink = {
+  title: '',
+  searchTag: 'lux:placeParts',
+  tab: 'places',
+}