+++ client/src/features/results/Paginate.tsx
@@ -166,7 +166,7 @@ const Paginate: React.FC<IPagination> = ({
               {paginationRange.length > 0 &&
                 paginationRange.map((pageNumber: string | number) => {
                   if (pageNumber === DOTS) {
-                    return <Pagination.Ellipsis disabled />
+                    return <Pagination.Ellipsis disabled key={pageNumber} />
                   }
                   return (
                     <Pagination.Item