+++ client/src/lib/util/hierarchyHelpers.ts
@@ -1,7 +1,10 @@
+import { Edge, Node } from 'reactflow'
+
 import IEntity from '../../types/data/IEntity'
 import ILinks from '../../types/data/ILinks'
 import ConceptParser from '../parse/data/ConceptParser'
 import SetParser from '../parse/data/SetParser'
+import { ISearchResults } from '../../types/ISearchResults'
 
 export const isInHierarchy = (uri: string, ancestors: Array<string>): boolean =>
   ancestors.includes(uri)
@@ -107,3 +110,95 @@ export const getParentData = (
   }
   return data[0]
 }
+
+const DEFAULT_POSITION = { x: 0, y: 0 }
+
+/**
+ * Parses the data and returns nodes for react flow hierarchy
+ * @param {Array<any>} data the current entity to parse
+ * @returns {Array<Node>}
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const getParentNodes = (data: Array<string>): Array<Node> =>
+  data.map((id) => ({
+    id,
+    type: 'hierarchyNode',
+    data: {
+      label: id, // this will be a RecordLink
+    },
+    position: DEFAULT_POSITION,
+  }))
+
+/**
+ * Parses the data and returns nodes for react flow hierarchy
+ * @param {Array<any>} data the current entity to parse
+ * @returns {Array<Node>}
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const getChildNodes = (data: ISearchResults): Array<Node> => {
+  const { orderedItems } = data
+  if (orderedItems.length === 0) {
+    console.log('No children were returned with the provided search criteria.')
+
+    return []
+  }
+
+  return orderedItems.map((obj) => ({
+    id: obj.id,
+    type: 'hierarchyNode',
+    data: {
+      label: obj.id, // this will be a RecordLink
+    },
+    position: DEFAULT_POSITION,
+  }))
+}
+
+/**
+ * Parses the nodes to get the correct data
+ * @param {Array<Node>} nodes the current nodes
+ * @param {string} currentUuid the current entity uuid
+ * @returns {Array<Edge>}
+ */
+export const getParentEdges = (
+  nodes: Array<Node>,
+  currentUuid: string,
+): Array<Edge> =>
+  nodes.map((node, ind) => ({
+    id: `e${ind}-${node.id}`, // the uuid of the parent
+    source: node.id, // the source is to the left of the target so the source is the parent
+    targetHandle: 'a', // handle on the target, a will be on the left of current
+    target: currentUuid,
+    // type: 'hierarchyEdge',
+  }))
+
+/**
+ * Parses the nodes to get the correct data
+ * @param {Array<Node>} nodes the current nodes
+ * @param {string} currentUuid the current entity uuid
+ * @returns {Array<Edge>}
+ */
+export const getChildEdges = (
+  nodes: Array<Node>,
+  currentUuid: string,
+): Array<Edge> =>
+  nodes.map((node, ind) => ({
+    id: `e${ind}-${node.id}`, // the uuid of the parent
+    source: currentUuid, // the source is to the left of the target so the source is the parent
+    sourceHandle: 'b', // handle on the target, a will be on the left of current
+    target: node.id,
+    // type: 'hierarchyEdge',
+  }))
+
+/**
+ * Returns the react flow node for the current entity
+ * @param {string} currentUuid the current entity uuid
+ * @returns {Node}
+ */
+export const getDefaultNode = (currentUuid: string): Node => ({
+  id: currentUuid,
+  type: 'hierarchyNode',
+  data: {
+    label: currentUuid, // this will be a RecordLink
+  },
+  position: DEFAULT_POSITION,
+})