+++ client/src/test/unit/features/common/RecordLink.spec.tsx
@@ -26,7 +26,7 @@ describe('RecordLink', () => {
       getName.mockReturnValueOnce({
         data: mockEntity,
         isSuccess: true,
-        refetch(): void {
+        refetch(): any {
           throw new Error('Function not implemented.')
         },
       })
@@ -89,7 +89,7 @@ describe('RecordLink', () => {
         data: undefined,
         isSuccess: false,
         isError: true,
-        refetch(): void {
+        refetch(): any {
           throw new Error('Function not implemented.')
         },
       })